requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
pyyaml>=6.0.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dateutil>=2.8.2
pydantic>=2.0.0
